chore: Docs, tidy up, rename files
