+++ editor.planx.uk/src/pages/FlowEditor/components/Sidebar/Search/mocks/getDisplayDetailsForResult.ts
